# Make `unquoteIfQuoted` a generic function to avoid casting