--- a/decimal_bench_test.go
+++ b/decimal_bench_test.go
@@ -248,3 +248,15 @@ func BenchmarkDecimal_ExpTaylor(b *testing.B) {
 		_, _ = d.ExpTaylor(10)
 	}
 }
+
+func BenchmarkDecimal_UnmarshalJSON(b *testing.B) {
+	b.ResetTimer()
+
+	data := []byte(`100.0000`)
+
+	b.ReportAllocs()
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		_ = (&Decimal{}).UnmarshalJSON(data)
+	}
+}