Our organization heavily relies on this package for all kinds of decimal operations. We do a lot of operations related to JSON parsing. As a low-hanging fruit for optimizations, we can start by making `unquoteIfQuoted` a generic function.
This function was causing unnecessary allocations by doing `v := value.(type)` casting. Making this function generic seems to give some performance.
Benchmarks show there's about 16% speedup

Benchmarks
Before:
```
goos: darwin
goarch: arm64
pkg: github.com/advbet/xdecimal/v2
cpu: Apple M1 Pro
BenchmarkDecimal_UnmarshalJSON-8   	 8665200	       130.8 ns/op	      80 B/op	       5 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	 8990174	       131.7 ns/op	      80 B/op	       5 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	 9205892	       130.3 ns/op	      80 B/op	       5 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	 9095335	       130.3 ns/op	      80 B/op	       5 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	 9113773	       134.5 ns/op	      80 B/op	       5 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	 9128460	       131.4 ns/op	      80 B/op	       5 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	 8823780	       137.7 ns/op	      80 B/op	       5 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	 8578131	       154.7 ns/op	      80 B/op	       5 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	 8883667	       134.3 ns/op	      80 B/op	       5 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	 9244027	       133.1 ns/op	      80 B/op	       5 allocs/op
PASS
ok  	github.com/advbet/xdecimal/v2	13.669s
```

After:
```
goos: darwin
goarch: arm64
pkg: github.com/advbet/xdecimal/v2
cpu: Apple M1 Pro
BenchmarkDecimal_UnmarshalJSON-8   	10049608	       112.0 ns/op	      56 B/op	       4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	10732308	       113.8 ns/op	      56 B/op	       4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	10185679	       138.7 ns/op	      56 B/op	       4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	10644686	       110.5 ns/op	      56 B/op	       4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	10641768	       110.8 ns/op	      56 B/op	       4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	10746822	       110.5 ns/op	      56 B/op	       4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	10720071	       110.9 ns/op	      56 B/op	       4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	10639095	       110.9 ns/op	      56 B/op	       4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	10615526	       111.1 ns/op	      56 B/op	       4 allocs/op
BenchmarkDecimal_UnmarshalJSON-8   	10766797	       111.6 ns/op	      56 B/op	       4 allocs/op
PASS
ok  	github.com/advbet/xdecimal/v2	13.428s
```
Diff:
```
                        │ bench1.txt  │             bench2.txt              │
                        │   sec/op    │   sec/op     vs base                │
Decimal_UnmarshalJSON-8   132.4n ± 4%   111.0n ± 3%  -16.16% (p=0.001 n=10)

                        │ bench1.txt │             bench2.txt             │
                        │    B/op    │    B/op     vs base                │
Decimal_UnmarshalJSON-8   80.00 ± 0%   56.00 ± 0%  -30.00% (p=0.000 n=10)

                        │ bench1.txt │             bench2.txt             │
                        │ allocs/op  │ allocs/op   vs base                │
Decimal_UnmarshalJSON-8   5.000 ± 0%   4.000 ± 0%  -20.00% (p=0.000 n=10)
```