@wascript3r 2025-03-07
What do you think about bumping lib Go version to 1.18 in `go.mod` and using generics instead?

```go
func unquoteIfQuoted[T ~string | ~[]byte](v T) T {
	if len(v) > 2 && v[0] == '"' && v[len(v)-1] == '"' {
		return v[1 : len(v)-1]
	}
	return v
}
```

I think benchmarks are the same:

Before:
```
Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -bench ^BenchmarkDecimal_UnmarshalJSON$ github.com/advbet/xdecimal/v2

goos: darwin
goarch: arm64
pkg: github.com/advbet/xdecimal/v2
cpu: Apple M4 Pro
=== RUN   BenchmarkDecimal_UnmarshalJSON
BenchmarkDecimal_UnmarshalJSON
BenchmarkDecimal_UnmarshalJSON-12       12697600                96.08 ns/op           80 B/op          5 allocs/op
PASS
ok      github.com/advbet/xdecimal/v2   1.487s
```

After:
```
Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -bench ^BenchmarkDecimal_UnmarshalJSON$ github.com/advbet/xdecimal/v2

goos: darwin
goarch: arm64
pkg: github.com/advbet/xdecimal/v2
cpu: Apple M4 Pro
=== RUN   BenchmarkDecimal_UnmarshalJSON
BenchmarkDecimal_UnmarshalJSON
BenchmarkDecimal_UnmarshalJSON-12       14218682                76.04 ns/op           56 B/op          4 allocs/op
PASS
ok      github.com/advbet/xdecimal/v2   1.335s
```