@PauliusLozys 2025-03-17
@wascript3r updated to use generic function.